requirements.txt
azure-identity
Flask==2.3.2
openai==1.40.6
python-dotenv==1.0.1
azure-core==1.30.2
azure-search-documents==11.6.0b4
aiohttp==3.10.5
aiofiles==24.1.0
graphrag
promptflow
promptflow-tracing
fastapi
uvicorn
pandas
tiktoken
graphrag
pydantic
python-dotenv
asyncio
aiohttp